requirements.txt
absl-py==1.0.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
black==21.4b2
brotlipy==0.7.0
cachetools==5.0.0
certifi==2021.10.8
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.0.0
cryptography @ file:///tmp/build/80754af9/cryptography_1639414572950/work
cycler==0.11.0
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@a5f284560e73d14d1563acb6f19c1d0e04226c85
docbank-loader @ file:///home/sysadminkcdh/subset_selection/DocBankLoader
fonttools==4.33.3
future==0.18.2
fvcore==0.1.5.post20220414
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.44.0
hydra-core==1.1.2
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata==4.11.3
iopath==0.1.9
joblib==1.1.0
kiwisolver==1.4.2
llvmlite==0.38.0
Markdown==3.3.6
matplotlib==3.5.1
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
mypy-extensions==0.4.3
numba==0.55.1
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1649764630438/work
oauthlib==3.2.0
omegaconf==2.1.2
opencv-python==4.5.5.64
packaging==21.3
panda==0.3.1
pandas==1.4.2
pathspec==0.9.0
Pillow==9.0.1
portalocker==2.4.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydot==1.4.2
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.8
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.4.24
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
submodlib @ file:///home/sysadminkcdh/subset_selection/submodlib
tabulate==0.8.9
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
threadpoolctl==3.1.0
toml==0.10.2
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.64.0
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3 @ file:///opt/conda/conda-bld/urllib3_1650639997961/work
Werkzeug==2.1.2
yacs==0.1.8
zipp==3.8.0
